init: fix themed boot animation on encrypted devices
On a non-encrypted boot, data is mounted, SurfaceFlinger starts,
SurfaceFlinger starts boot animation, and everything works as expected.
On an encrypted boot, SurfaceFlinger starts, bootanimation starts, decryption
happens, and bootanimation continues without reloading because SurfaceFlinger is
not reloaded, and thus netiher is bootanimation.

This commit forces bootanimation to restart once /data is decrypted so that
it picks up the themed boot animation.

Change-Id: If429b2764051f161dfd3602cf29a9911077b70ed
Signed-off-by: Evan Anderson <evananderson@aospa.co>