debuggerd: Generate core-dumps for native userspace crashes
We aim to generate coredumps when certain criteria are met
on userdebug/eng builds. For instance, ART crashes, which
are difficult to reproduce to begin with.

For this, debuggerd needs to have capability to set resource
limits for the crashing process, with the help of prlimit().

More criteria can be added as required.

By default, the feature will be enabled on userdebug/eng builds;
however, it can be disabled by setting the property
"persist.force_coredump.disable" to 1.

Change-Id: I7b68afb868281d9c6f3df62da24b430b9f64f6ea