init.rc: Fix potential task-throttling for regular tasks
Linux recommends the configuration where a second
is divided into 950ms for RT tasks and 50ms for
regular tasks.
This limit where regular tasks are guaranteed to
get 50ms of cpu time is meant to be a security
mechanism for cases where RT tasks misbehave
and take up too much cputime.

However in Android where we have multiple userspace
processes running as RT, this limit is being reached
constantly.

So now it becomes obvious that 50ms for the regular
tasks (SCHED_OTHER) is not enough.

To improve the situation, this patch changes the setup
and provides 900ms for RT tasks and 100ms for regular
tasks.
The damage caused by the extended throttling of RT tasks
is negligible.

Change-Id: Ic84ac7efaad0302e70434e4cb411353e15cbf6f5
Signed-off-by: Alex Naidis <alex.naidis@linux.com>